requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
schedule==1.2.1
selenium==4.18.1
webdriver-manager==4.0.1
python-telegram-bot==20.8